+++ artifacts/kuhni-na-zakaz/.next/static/chunks/webpack.js
@@ -190,7 +190,7 @@
 /******/ 	
 /******/ 	/* webpack/runtime/getFullHash */
 /******/ 	(() => {
-/******/ 		__webpack_require__.h = () => ("796f26dde552037f")
+/******/ 		__webpack_require__.h = () => ("f01be9d6af752dbc")
 /******/ 	})();
 /******/ 	
 /******/ 	/* webpack/runtime/global */